split multisig into modules
