+++ ethereum-fee-prepay/src/aggregator_result.rs
@@ -1,5 +1,4 @@
 elrond_wasm::imports!();
-use elrond_wasm::sc_error;
 
 use transaction::*;
