+++ multisig/src/multisig_general.rs
@@ -1,0 +1,180 @@
+elrond_wasm::imports!();
+
+use crate::action::Action;
+use crate::user_role::UserRole;
+
+#[elrond_wasm_derive::module]
+pub trait MultisigGeneralModule: crate::util::UtilModule + crate::storage::StorageModule {
+    /// Used by board members to sign actions.
+    #[endpoint]
+    fn sign(&self, action_id: usize) -> SCResult<()> {
+        require!(
+            !self.action_mapper().item_is_empty_unchecked(action_id),
+            "action does not exist"
+        );
+
+        let caller_address = self.blockchain().get_caller();
+        let caller_id = self.user_mapper().get_user_id(&caller_address);
+        let caller_role = self.get_user_id_to_role(caller_id);
+        require!(caller_role.can_sign(), "only board members can sign");
+        require!(self.has_enough_stake(&caller_address), "not enough stake");
+
+        self.action_signer_ids(action_id).update(|signer_ids| {
+            if !signer_ids.contains(&caller_id) {
+                signer_ids.push(caller_id);
+            }
+        });
+
+        Ok(())
+    }
+
+    /// Board members can withdraw their signatures if they no longer desire for the action to be executed.
+    /// Actions that are left with no valid signatures can be then deleted to free up storage.
+    #[endpoint]
+    fn unsign(&self, action_id: usize) -> SCResult<()> {
+        require!(
+            !self.action_mapper().item_is_empty_unchecked(action_id),
+            "action does not exist"
+        );
+
+        let caller_address = self.blockchain().get_caller();
+        let caller_id = self.user_mapper().get_user_id(&caller_address);
+        let caller_role = self.get_user_id_to_role(caller_id);
+        require!(caller_role.can_sign(), "only board members can un-sign");
+        require!(self.has_enough_stake(&caller_address), "not enough stake");
+
+        self.action_signer_ids(action_id).update(|signer_ids| {
+            if let Some(signer_pos) = signer_ids
+                .iter()
+                .position(|&signer_id| signer_id == caller_id)
+            {
+                // since we don't care about the order,
+                // it is ok to call swap_remove, which is O(1)
+                signer_ids.swap_remove(signer_pos);
+            }
+        });
+
+        Ok(())
+    }
+
+    /// Clears storage pertaining to an action that is no longer supposed to be executed.
+    /// Any signatures that the action received must first be removed, via `unsign`.
+    /// Otherwise this endpoint would be prone to abuse.
+    #[endpoint(discardAction)]
+    fn discard_action(&self, action_id: usize) -> SCResult<()> {
+        let caller_address = self.blockchain().get_caller();
+        let caller_id = self.user_mapper().get_user_id(&caller_address);
+        let caller_role = self.get_user_id_to_role(caller_id);
+        require!(
+            caller_role.can_discard_action(),
+            "only board members and proposers can discard actions"
+        );
+        require!(
+            self.get_action_valid_signer_count(action_id) == 0,
+            "cannot discard action with valid signatures"
+        );
+
+        self.clear_action(action_id);
+        Ok(())
+    }
+
+    /// Initiates board member addition process.
+    /// Can also be used to promote a proposer to board member.
+    #[endpoint(proposeAddBoardMember)]
+    fn propose_add_board_member(&self, board_member_address: Address) -> SCResult<usize> {
+        self.propose_action(Action::AddBoardMember(board_member_address))
+    }
+
+    /// Initiates proposer addition process..
+    /// Can also be used to demote a board member to proposer.
+    #[endpoint(proposeAddProposer)]
+    fn propose_add_proposer(&self, proposer_address: Address) -> SCResult<usize> {
+        self.propose_action(Action::AddProposer(proposer_address))
+    }
+
+    /// Removes user regardless of whether it is a board member or proposer.
+    #[endpoint(proposeRemoveUser)]
+    fn propose_remove_user(&self, user_address: Address) -> SCResult<usize> {
+        self.propose_action(Action::RemoveUser(user_address))
+    }
+
+    #[endpoint(proposeChangeQuorum)]
+    fn propose_change_quorum(&self, new_quorum: usize) -> SCResult<usize> {
+        self.propose_action(Action::ChangeQuorum(new_quorum))
+    }
+
+    #[endpoint(proposeSlashUser)]
+    fn propose_slash_user(&self, user_address: Address) -> SCResult<usize> {
+        self.require_caller_owner()?;
+        require!(
+            self.user_role(&user_address) == UserRole::BoardMember,
+            "can only slash board members"
+        );
+
+        self.propose_action(Action::SlashUser(user_address))
+    }
+
+    fn propose_action(&self, action: Action<Self::BigUint>) -> SCResult<usize> {
+        let caller_address = self.blockchain().get_caller();
+        let caller_id = self.user_mapper().get_user_id(&caller_address);
+        let caller_role = self.get_user_id_to_role(caller_id);
+        require!(
+            caller_role.can_propose(),
+            "only board members and proposers can propose"
+        );
+
+        let action_id = self.action_mapper().push(&action);
+        if caller_role.can_sign() {
+            // also sign
+            // since the action is newly created, the caller can be the only signer
+            self.action_signer_ids(action_id).set(&[caller_id].to_vec());
+        }
+
+        Ok(action_id)
+    }
+
+    fn clear_action(&self, action_id: usize) {
+        self.action_mapper().clear_entry_unchecked(action_id);
+        self.action_signer_ids(action_id).clear();
+    }
+
+    /// Can be used to:
+    /// - create new user (board member / proposer)
+    /// - remove user (board member / proposer)
+    /// - reactivate removed user
+    /// - convert between board member and proposer
+    /// Will keep the board size and proposer count in sync.
+    fn change_user_role(&self, user_address: Address, new_role: UserRole) {
+        let user_id = self.user_mapper().get_or_create_user(&user_address);
+        let old_role = if user_id == 0 {
+            UserRole::None
+        } else {
+            self.get_user_id_to_role(user_id)
+        };
+        self.set_user_id_to_role(user_id, new_role);
+
+        // update board size
+        #[allow(clippy::collapsible_else_if)]
+        if old_role == UserRole::BoardMember {
+            if new_role != UserRole::BoardMember {
+                self.num_board_members().update(|value| *value -= 1);
+            }
+        } else {
+            if new_role == UserRole::BoardMember {
+                self.num_board_members().update(|value| *value += 1);
+            }
+        }
+
+        // update num_proposers
+        #[allow(clippy::collapsible_else_if)]
+        if old_role == UserRole::Proposer {
+            if new_role != UserRole::Proposer {
+                self.num_proposers().update(|value| *value -= 1);
+            }
+        } else {
+            if new_role == UserRole::Proposer {
+                self.num_proposers().update(|value| *value += 1);
+            }
+        }
+    }
+}
